Fix links and do minor cleanup in the Julia style guide
Many of the links to the Julia documentation are currently 404s, so this
adjusts them to point to the right places.

A section that's only relevant to very old Julia code (0.3 and earlier)
has been removed.

"Old style" parametric syntax, which is formally deprecated in 0.7 and
will be repurposed in 1.0+, has been converted to 0.6-style `where`
clauses. Any code written now that uses the old style will be broken
once 1.0 is released.

A note about the BenchmarkTools package has been added to the
Performance and Optimization section, as it partially addresses one of
the concerns noted therein.